python/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["s213610"]

[tool.poetry.dependencies]
python = "3.10"
grpcio = "1.50.0"
grpcio-tools = "1.50.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
